requirements.txt
absl-py==0.8.1
alabaster==0.7.12
astor==0.8.0
atari-py==0.2.6
attrdict==2.0.1
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
beautifultable==0.8.0
bleach==3.1.5
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
docutils==0.15.2
entrypoints==0.3
fasteners==0.15
future==0.18.2
gast==0.3.2
glfw==1.8.5
glob2==0.7
google-pasta==0.1.8
grpcio==1.25.0
-e git+https://github.com/erick-alv/HGG-extended-master.git@a7ec797f9f046384fc5cc7415352200edd598dd5#egg=gym&subdirectory=gym
h5py==2.10.0
hickle==4.0.0
idna==2.8
ImageHash==4.0
imageio==2.6.1
imagesize==1.1.0
importlib-metadata==1.2.0
ipdb==0.12.3
ipykernel==5.3.0
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lockfile==0.12.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
memory-profiler==0.57.0
mistune==0.8.4
monotonic==1.5
more-itertools==8.0.2
mpi4py==3.0.3
-e git+https://github.com/openai/mujoco-py.git@1fe312b09ae7365f0dd9d4d0e453f8da59fae0bf#egg=mujoco_py
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numpy==1.16.4
numpydoc==0.9.1
opencv-python==4.1.2.30
packaging==19.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.1
progressbar2==3.47.0
prometheus-client==0.8.0
prompt-toolkit==2.0.10
protobuf==3.11.1
psutil==5.7.2
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
pyglet==1.3.2
Pygments==2.5.2
pyparsing==2.4.5
pyrsistent==0.16.0
pytest==5.3.1
pytest-instafail==0.3.0
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2019.3
PyWavelets==1.1.1
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.22.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.3.3
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
snowballstemmer==2.0.0
Sphinx==2.2.2
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.6.3
torch==1.5.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.40.1
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
zmq==0.0.0

